renamed utilityConstants to formatting and moved colors to their own file
