+++ src/client/components/chart/index.js
@@ -4,7 +4,7 @@ import Chart from './Chart';
 import { getSelectedStockValueForKey } from '../../../utils/stateGetters';
 import {
   CHART_META_DATA
-} from '../../../constants/utilityConstants';
+} from '../../../constants/formatting';
 
 const getStockDataKey = timePeriodIndex =>
   CHART_META_DATA[timePeriodIndex].stockDataKey;
