+++ src/client/components/searchLoader/SearchLoader.jsx
@@ -2,10 +2,8 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { BeatLoader, ScaleLoader } from 'react-spinners';
 
-import {
-  SEARCH_STATUS_REGION_HEIGHT,
-  THEME_COLOR_DARK1
-} from '../../../constants/utilityConstants';
+import { SEARCH_STATUS_REGION_HEIGHT } from '../../../constants/formatting';
+import { THEME_COLOR_DARK1 } from '../../../constants/colors';
 
 const SearchLoader = ({ showingCachedStock }) => {
   const loaderProps = {
