+++ src/client/reducers/chartTimePeriodIndex.js
@@ -2,7 +2,7 @@ import {
   SET_CHART_TO_DEFAULT_TIME_PERIOD,
   UPDATE_CHART_TIME_PERIOD_INDEX
 } from '../actions';
-import { INDEX_ONE_YEAR } from '../../constants/utilityConstants';
+import { INDEX_ONE_YEAR } from '../../constants/formatting';
 
 const DEFAULT_TIME_PERIOD_INDEX = INDEX_ONE_YEAR;
