+++ src/constants/colors.js
@@ -1,0 +1,4 @@
+export const CHART_LINE_COLOR = 'red';
+
+export const THEME_COLOR_DARK1 = '#34414D';
+export const THEME_COLOR_MEDIUM1 = '#5998ff';
