+++ src/utils/formatting/stringFormatting.js
@@ -1,6 +1,6 @@
 import numeral from 'numeral';
 
-import { NUMBER_FORMAT_DEFAULT } from '../../constants/utilityConstants';
+import { NUMBER_FORMAT_DEFAULT } from '../../constants/formatting';
 import { BLANK_FIELD } from '../../constants/userFacingStrings';
 import { getSelectedStockValueForKey } from '../stateGetters';
